separating web part as microservice | started work on message_analyzer microservice
