+++ message_analyzer_microservice/src/main/java/com/jejutic/message_analyzer/MessageAnalyzerApplication.java
@@ -1,0 +1,13 @@
+package com.jejutic.message_analyzer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MessageAnalyzerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MessageAnalyzerApplication.class, args);
+	}
+
+}
