+++ message_analyzer_microservice/src/main/java/com/jejutic/message_analyzer/controller/AnalyzeRequestController.java
@@ -1,0 +1,17 @@
+package com.jejutic.message_analyzer.controller;
+
+import com.jejutic.message_analyzer.dto.Message;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1")
+public class AnalyzeRequestController {
+
+    @PostMapping("/analyze")
+    public String randomEnding(@RequestBody Message message) {
+        return "RandomEnding";
+    }
+}
