+++ message_analyzer_microservice/src/main/java/com/jejutic/message_analyzer/dto/Message.java
@@ -1,0 +1,11 @@
+package com.jejutic.message_analyzer.dto;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Message {
+
+    private String text;
+}
