+++ message_analyzer_microservice/src/main/resources/application.yaml
@@ -1,0 +1,2 @@
+server:
+  port: 8090
