+++ message_analyzer_microservice/src/test/java/com/jejutic/message_analyzer/MessageAnalyzerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.jejutic.message_analyzer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MessageAnalyzerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
