--- src/main/java/com/jejutic/social/SocialApplication.java
@@ -1,13 +1,0 @@
-package com.jejutic.social;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SocialApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SocialApplication.class, args);
-	}
-
-}
