--- src/main/java/com/jejutic/social/ui/LoginPage.java
@@ -1,31 +1,0 @@
-package com.jejutic.social.ui;
-
-import com.jejutic.social.ui.data_objects.UserInfo;
-import jakarta.validation.Valid;
-import org.springframework.stereotype.Controller;
-import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.*;
-
-@Controller
-@RequestMapping("/login")
-@SessionAttributes("userInfo")
-public class LoginPage {
-
-    @ModelAttribute("userInfo")
-    public UserInfo userInfo() {
-        return new UserInfo();
-    }
-
-    @GetMapping
-    public String login() {
-        return "login";
-    }
-
-    @PostMapping
-    public String login(@Valid UserInfo userInfo, Errors errors) {
-        if (errors.hasErrors()) {
-            return "redirect:/login";
-        }
-        return "redirect:/";
-    }
-}
