--- src/main/java/com/jejutic/social/ui/data_objects/Message.java
@@ -1,12 +1,0 @@
-package com.jejutic.social.ui.data_objects;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Message {
-
-    private String text;
-    private String author;
-}
