--- src/main/java/com/jejutic/social/ui/data_objects/UserInfo.java
@@ -1,13 +1,0 @@
-package com.jejutic.social.ui.data_objects;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-
-@Data
-public class UserInfo {
-
-    @NotNull
-    @Size(min=1, max=20)
-    private String nickname;
-}
