+++ web_service/src/main/java/com/jejtuic/web_service/SocialApplication.java
@@ -1,0 +1,13 @@
+package com.jejtuic.web_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SocialApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SocialApplication.class, args);
+	}
+
+}
