+++ web_service/src/main/java/com/jejtuic/web_service/configuration/ApplicationConfiguration.java
@@ -1,0 +1,15 @@
+package com.jejtuic.web_service.configuration;
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class ApplicationConfiguration {
+
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+}
