+++ web_service/src/main/java/com/jejtuic/web_service/database/data_objects/Message.java
@@ -1,0 +1,13 @@
+package com.jejtuic.web_service.database.data_objects;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Message {
+
+    private Long id;
+    private String text;
+    private String author;
+}
