+++ web_service/src/main/java/com/jejtuic/web_service/ui/data_objects/Message.java
@@ -1,0 +1,12 @@
+package com.jejtuic.web_service.ui.data_objects;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Message {
+
+    private String text;
+    private String author;
+}
