+++ web_service/src/main/java/com/jejtuic/web_service/ui/data_objects/UserInfo.java
@@ -1,0 +1,13 @@
+package com.jejtuic.web_service.ui.data_objects;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+
+@Data
+public class UserInfo {
+
+    @NotNull
+    @Size(min=1, max=20)
+    private String nickname;
+}
