+++ web_service/src/main/resources/templates/main_page.html
